--- src/main/java/yapp/buddycon/common/domain/AuthMember.java
@@ -1,7 +1,0 @@
-package yapp.buddycon.common.domain;
-
-public record AuthMember(
-  Long id
-) {
-
-}
