+++ src/main/java/yapp/buddycon/web/auth/adapter/in/response/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.web.member.auth.adapter.in.response;
+package yapp.buddycon.web.auth.adapter.in.response;
 
 public record TokenResponseDto(
   String grantType,
