+++ src/main/java/yapp/buddycon/web/auth/application/port/in/OAuthMemberInfo.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.web.member.auth.domain;
+package yapp.buddycon.web.auth.application.port.in;
 
 public record OAuthMemberInfo(
   Long id,
