+++ src/main/java/yapp/buddycon/web/auth/application/service/TokenDecoder.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.web.member.auth.application.service;
+package yapp.buddycon.web.auth.application.service;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
